Revert "Issue #4220: Modified EmptyBlockCheckTest.java and moved its input files to the emptyblock subdirectory"
This reverts commit 2036e7a4abceddad23311f4fa4bc61460072756a.
