+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/blocks/InputSemantic.java
@@ -2,7 +2,7 @@
 // Test case file for checkstyle.
 // Created: 2001
 ////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks.blocks.emptyblock;
+package com.puppycrawl.tools.checkstyle.checks.blocks;
 
 import java.io.*; // star import for instantiation tests
 import java.awt.Dimension; // explicit import for instantiation tests
@@ -12,7 +12,7 @@ import java.awt.Color;
  * Test case for detecting empty block statements.
  * @author Lars Kühne
  **/
-class InputEmptyBlockDefault
+class InputSemantic
 {
     static {
         Boolean x = new Boolean(true);
